+++ lib/EasyPost/Service/ShipmentService.php
@@ -30,12 +30,7 @@ class ShipmentService extends BaseService
      */
     public function all($params = null)
     {
-        self::validate($params);
-        $response = Requestor::request($this->client, 'get', '/shipments', $params);
-        $response['purchased'] = $params['purchased'] ?? null;
-        $response['include_children'] = $params['include_children'] ?? null;
-
-        return InternalUtil::convertToEasyPostObject($this->client, $response);
+        return self::allResources(self::serviceModelClassName(self::class), $params);
     }
 
     /**
@@ -47,16 +42,7 @@ class ShipmentService extends BaseService
      */
     public function getNextPage($shipments, $pageSize = null)
     {
-        $params = [];
-
-        if (isset($shipments->purchased)) {
-            $params['purchased'] = $shipments->purchased;
-        }
-
-        if (isset($shipments->include_children)) {
-            $params['include_children'] = $shipments->include_children;
-        }
-        return $this->getNextPageResources(self::serviceModelClassName(self::class), $shipments, $pageSize, $params);
+        return $this->getNextPageResources(self::serviceModelClassName(self::class), $shipments, $pageSize);
     }
 
     /**