+++ test/EasyPost/PickupTest.php
@@ -3,10 +3,10 @@
 namespace EasyPost\Test;
 
 use EasyPost\EasyPostClient;
-use EasyPost\Exception\General\FilteringException;
 use EasyPost\Exception\General\EndOfPaginationException;
-use Exception;
+use EasyPost\Exception\General\FilteringException;
 use EasyPost\Pickup;
+use Exception;
 
 class PickupTest extends \PHPUnit\Framework\TestCase
 {
@@ -102,11 +102,11 @@ class PickupTest extends \PHPUnit\Framework\TestCase
             $secondIdOfSecondPage = $nextPage['pickups'][0]->id;
 
             $this->assertNotEquals($firstIdOfFirstPage, $secondIdOfSecondPage);
-        } catch (Exception $error) {
-            if (!($error instanceof EndOfPaginationException)) {
-                throw new Exception('Test failed intentionally');
-            }
+        } catch (EndOfPaginationException $error) {
+            // There's no second page, that's not a failure
             $this->assertTrue(true);
+        } catch (Exception $error) {
+            throw $error;
         }
     }