+++ test/EasyPost/RefundTest.php
@@ -86,11 +86,11 @@ class RefundTest extends \PHPUnit\Framework\TestCase
             $secondIdOfSecondPage = $nextPage['refunds'][0]->id;
 
             $this->assertNotEquals($firstIdOfFirstPage, $secondIdOfSecondPage);
-        } catch (Exception $error) {
-            if (!($error instanceof EndOfPaginationException)) {
-                throw new Exception('Test failed intentionally');
-            }
+        } catch (EndOfPaginationException $error) {
+            // There's no second page, that's not a failure
             $this->assertTrue(true);
+        } catch (Exception $error) {
+            throw $error;
         }
     }