+++ test/EasyPost/TrackerTest.php
@@ -112,11 +112,11 @@ class TrackerTest extends \PHPUnit\Framework\TestCase
             $secondIdOfSecondPage = $nextPage['trackers'][0]->id;
 
             $this->assertNotEquals($firstIdOfFirstPage, $secondIdOfSecondPage);
-        } catch (Exception $error) {
-            if (!($error instanceof EndOfPaginationException)) {
-                throw new Exception('Test failed intentionally');
-            }
+        } catch (EndOfPaginationException $error) {
+            // There's no second page, that's not a failure
             $this->assertTrue(true);
+        } catch (Exception $error) {
+            throw $error;
         }
     }