+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Next Major Release
 
+- Drops support for PHP 7.4
+- Adds support for PHP 8.3
 - Removed `withCarbonOffset` parameter from `create`, `buy`, and `regenerateRates` functions of the Shipment service as EasyPost now offers Carbon Neutral shipments by default for free
 - Fixes a bug where the original filtering criteria of `all` calls wasn't passed along to `getNextPage` calls. Now, these are persisted via a `_params` key on response objects locally